% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bp_algorithms.R
\name{bp_algorithm}
\alias{bp_algorithm}
\title{Breakpoint Algorithm}
\usage{
bp_algorithm(
  .data,
  algorithm,
  .x,
  .y,
  vo2 = "vo2",
  vco2 = "vco2",
  ve = "ve",
  time = "time"
)
}
\arguments{
\item{.data}{Gas exchange data.}

\item{algorithm}{The specific breakpoint algorithm you want to use to find a given threshold. Choices include \code{dmax}, \code{dmax_mod}, \code{jm}, \code{orr}, \code{v-slope}, \code{v_slope_simple}, and \code{splines}.}

\item{.x}{The x-axis variable}

\item{.y}{the y-axis variable}
}
\value{

}
\description{
Breakpoint Algorithm
}
\examples{
# Come up with an example later

}
