% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/breakpoints.R
\name{breakpoint}
\alias{breakpoint}
\title{Find breakpoints in gas exchange data}
\usage{
breakpoint(
  .data,
  method = NULL,
  algorithm_aert = NULL,
  x_aert = NULL,
  y_aert = NULL,
  algorithm_rc = NULL,
  x_rc = NULL,
  y_rc = NULL,
  vo2_col = "vo2",
  cco2_col = "vco2",
  ve_col = "ve",
  time_col = "time"
)
}
\arguments{
\item{.data}{Gas exchange data}

\item{method}{}

\item{.x}{Variable to use on the x axis, usually time, VO2, or VCO2.}

\item{.y}{Variable to use on the x axis}
}
\value{

}
\description{
Find breakpoints in gas exchange data
}
\details{
There are multiple ways to find a breakpoint. The first is by specifying a commonly used \code{method} and the appropriate \code{algorithm} and \code{.x}, and \code{y}, variables will be selected for you. For example, entering 'v-slope' for \code{method} will pre-select 'v-slope' for \code{algorithm}, 'vo2' for .x, and 'vco2' for .y. If you enter a specific \code{method}, you can override the default \code{algorithm}, \code{.x}, and \code{.y}.

Since there aren't common names for every way to find a breakpoint, you can also leave \code{method} as \code{NULL} and specify the \code{algorithm}, \code{.x}, and \code{.y}.
}
\examples{
# TODO write an an example

}
\references{
Beaver, W. L., Wasserman, K., & Whipp, B. J. (1986). A new method for detecting anaerobic threshold by gas exchange. Journal of Applied Physiology, 121(6), 2020–2027.
}
