% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/d1_crossing.R
\name{d1_crossing}
\alias{d1_crossing}
\title{Find the gas exchange threshold (GET) (~1st ventilatory threshold (VT1)) using the 1st derivative crossing point}
\usage{
d1_crossing(
  .data,
  .x,
  .y,
  degree = NULL,
  vo2 = "vo2",
  vco2 = "vco2",
  ve = "ve",
  time = "time",
  alpha_linearity = 0.05,
  pos_change = TRUE,
  bp = "vt1",
  ...
)
}
\arguments{
\item{.data}{Gas exchange data frame or tibble.}

\item{.x}{The x-axis variable.}

\item{.y}{the y-axis variable.}

\item{degree}{The polynomial degree the function should fit to the curve. If left \code{NULL} this function finds the best fit.}

\item{vo2}{Name of the \code{vo2} variable}

\item{vco2}{Name of the \code{vco2} variable}

\item{ve}{Name of the \code{ve} variable}

\item{time}{Name of the \code{time} variable}

\item{alpha_linearity}{Significance value to determine if a piecewise model explains significantly reduces the residual sums of squares more than a simpler model.}

\item{pos_change}{Do you expect the change in slope to be positive (default) or negative? If a two-line regression explains significantly reduces the sum square error but the change in slope does not match the expected underlying physiology, the breakpoint will be classified as indeterminate.}

\item{bp}{Is this the first (\code{vt1}) or the second (\code{vt2}) breakpoint? This method is specifically designed for finding VT1 (GET)}

\item{...}{Dot dot dot mostly allows this function to work properly if breakpoint() passes arguments that is not strictly needed by this function.}
}
\value{
A list include breakpoint data, best-fit and related functions, and a plot.
}
\description{
Use polynomial regression to find the gas exchange threshold (GET) (practically equivalent to the first ventilatory threshold (VT1)). This method by Wisén and Wohlfart (2004) is similar to the V-slope method by Beaver et al. (1986) in that it finds when the rate of CO2 production outpaces the rate of O2 production. In contrast to piecewise regression in the original V-slope method, this uses the first derivative of VO2 and VCO2 vs. time. The threshold is when dVCO2/dt surpasses dVO2/dt for the final time.
}
\examples{

# TODO write an example

}
\references{
Beaver, W. L., Wasserman, K. A. R. L. M. A. N., & Whipp, B. J. (1986). A new method for detecting anaerobic threshold by gas exchange. Journal of applied physiology, 60(6), 2020-2027.
Wisén, A. G., & Wohlfart, B. (2004). A refined technique for determining the respiratory gas exchange responses to anaerobic metabolism during progressive exercise–repeatability in a group of healthy men. Clinical physiology and functional imaging, 24(1), 1-9.
}
