% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/d1_poly_reg_maxima.R
\name{d1_poly_reg_maxima}
\alias{d1_poly_reg_maxima}
\title{Find a breakpoint using the maxima in the 1st derivative}
\usage{
d1_poly_reg_maxima(
  .data,
  .x,
  .y,
  bp,
  degree = 5,
  vo2 = "vo2",
  vco2 = "vco2",
  ve = "ve",
  time = "time",
  alpha_linearity = 0.05,
  make_plot = TRUE,
  show_plot = FALSE
)
}
\arguments{
\item{.data}{Gas exchange data frame or tibble.}

\item{.x}{The x-axis variable.}

\item{.y}{the y-axis variable.}

\item{bp}{Is this the first (\code{vt1}) or the second (\code{vt2}) breakpoint?}

\item{degree}{The polynomial degree the function should fit to the curve. If left \code{NULL} this function finds the best fit.}

\item{vo2}{Name of the \code{vo2} variable}

\item{vco2}{Name of the \code{vco2} variable}

\item{ve}{Name of the \code{ve} variable}

\item{time}{Name of the \code{time} variable}

\item{alpha_linearity}{Significance value to determine if a piecewise model explains significantly reduces the residual sums of squares more than a simpler model.}

\item{make_plot}{Make plot showing the breakpoint?}

\item{show_plot}{Show the plot showing the breakpoint? Default = `FALSE`.}
}
\value{
A slice of the original data frame at the threshold index with a new `algorithm` column.
}
\description{
Use polynomial regression of increasing order to find the best-fit line for the data, take the 1st derivative, and find the highest maxima of the 1st derivative. This method is traditionally used with ventilatory equivalents (VE/VO2 or VE/VCO2), PetO2, or PetCO2 vs. time or vs. VO2. This method was specifically implemented by Santos et al. (2004).
}
\examples{
# TODO write an example

}
\references{
Santos, E. L., & Giannella-Neto, A. (2004). Comparison of computerized methods for detecting the ventilatory thresholds. European journal of applied physiology, 93, 315-324.
}
