% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/d2_inflection.R
\name{d2_inflection}
\alias{d2_inflection}
\title{Second derivative inflection}
\usage{
d2_inflection(
  .data,
  .x,
  .y,
  bp,
  degree = 5,
  vo2 = "vo2",
  vco2 = "vco2",
  ve = "ve",
  time = "time",
  alpha_linearity = 0.05
)
}
\arguments{
\item{.data}{Gas exchange data frame or tibble.}

\item{.x}{The x-axis variable.}

\item{.y}{the y-axis variable.}

\item{bp}{Is this the first (\code{vt1}) or the second (\code{vt2}) breakpoint?}

\item{degree}{The polynomial degree the function should fit to the curve. If left \code{NULL} this function finds the best fit.}

\item{vo2}{Name of the \code{vo2} variable}

\item{vco2}{Name of the \code{vco2} variable}

\item{ve}{Name of the \code{ve} variable}

\item{time}{Name of the \code{time} variable}

\item{alpha_linearity}{Significance value to determine if a piecewise model explains significantly reduces the residual sums of squares more than a simpler model.}
}
\value{
A data frame at the threshold, with a few columns added to note the algorithm used.
}
\description{
Find a breakpoint in ventilatory data based on the 2nd derivative inflection point after Santos et al. (2004). In their paper, they use pulmonary ventilation (VE) vs. VO2 to find the first ventilatory (VT1) / gas exchange threshold (GET) and VE vs. VCO2 to find the respiratory compensation point (RCP) / second ventilatory threshold (VT2). It is generally suitable to also use VE vs. time to find VT2. One can truncate the test and rerun VE vs. time to find VT1.
}
\details{
To mimic the Santos et al. (2004) paper, this fits a 5th-order polynomial regression to the VE vs. VO2 or VCO2 data. Users can set `degree = NULL` to iteratively increase the polynomial if the likelihood ratio test is significant. After establishing the polynomial, we find the real roots of the 2nd derivative within the given range of the `.x` variable to find the inflection points. We then filter by concavity changes that go from up to down because this matches the expected shape of the pulmonary ventilation curves.

We are unsure how to handle multiple inflection points at this time, but this is unlikely to be an issue if the polynomial degree is kept equal to 5.
}
\examples{
# TODO write an example

}
\references{
Santos, E. L., & Giannella-Neto, A. (2004). Comparison of computerized methods for detecting the ventilatory thresholds. European journal of applied physiology, 93, 315-324.
}
