% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/d2_poly_reg_maxima.R
\name{d2_poly_reg_maxima}
\alias{d2_poly_reg_maxima}
\title{Find a breakpoint using the maxima in the 2nd derivative}
\usage{
d2_poly_reg_maxima(
  .data,
  .x,
  .y,
  bp,
  degree = NULL,
  vo2 = "vo2",
  vco2 = "vco2",
  ve = "ve",
  time = "time",
  alpha_linearity = 0.05
)
}
\arguments{
\item{.data}{Gas exchange data frame or tibble.}

\item{.x}{The x-axis variable.}

\item{.y}{the y-axis variable.}

\item{bp}{Is this the first (\code{vt1}) or the second (\code{vt2}) breakpoint?}

\item{degree}{The polynomial degree the function should fit to the curve. If left \code{NULL} this function finds the best fit.}

\item{vo2}{Name of the \code{vo2} variable}

\item{vco2}{Name of the \code{vco2} variable}

\item{ve}{Name of the \code{ve} variable}

\item{time}{Name of the \code{time} variable}

\item{alpha_linearity}{Significance value to determine if a piecewise model explains significantly reduces the residual sums of squares more than a simpler model.}
}
\value{
A slice of the original data frame at the threshold index with a new `algorithm` column.
}
\description{
Use polynomial regression of increasing order to find the best-fit line for the data, take the second derivative, and find the highest maxima of the second derivative. This method is traditionally used with ventilatory equivalents (VE/VO2 or VE/VCO2), PetO2, or PetCO2 vs. time or vs. VO2.
}
\details{
Note, polynomial regressions are usually inferior to regression splines or to smoothing splines because polynomial regressions because polynomial regressions can overfit the data more easily. However, we offer this method in case users want to reproduce previous works and while we develop this package.

Unlike several published works, this method iteratively finds the best-fit polynomial regression by increasing the polynomial order and using the likelihood ratio test with the \code{anova} function. When the likelihood ratio test does not find a statistically significant different difference between the previous and the newest polynomial, the function uses the previous polynomial. However, uses can specify the polynomial \code{degree} if desired.

#' @details
Our implementation is similar to two studies by Wisén and Wohlfart et al. in that this uses a polynomial regression. However, they only found the \emph{first} derivative of the VO2 and VCO2 vs. time curves and denote the threshold as their crossing point. Using ventilatory equivalents vs. time, they considered the threshold as when the \emph{first} derivative of each ventilatory equivalent increased above zero (for the last time). In contrast, the majority of other literature we are aware of that uses regression splines or smoothing splines finds the \emph{second} derivative. This function finds the second derivative by default to match the majority of other literature.

Taking either the first or the second derivative are both valid approaches, but we prefer the reasoning behind taking the second derivative. After crossing a threshold, ventilation increases out of proportion to VO2 (VT1) or to VCO2 (VT2). After the threshold, the rate of increase is faster. Put another way, the slope after the breakpoint is higher than the slope before the breakpoint. However, since we are more interested in when that \emph{slope changes}, we want to know the slope of the slope, i.e., the acceleration, or the second derivative.

Following the logic of Leo et al. (2017), this method then finds the local maxima in the acceleration because this shows when the slope is changing most rapidly. The idea is that above and below each threshold, our body is in two different states, and by finding the where slope changes most rapidly, we are locating that tipping point. Although other research shows that "thresholds" are \emph{not} tipping points but rather \emph{transitions} with gray zones between them (Ozkaya et al., 2022), it is nevertheless convenient to denote the threshold with a single point.

To implement a method similar to Leo et al. (2017), this function actually takes four derivatives of the best-fit polynomial. We find the roots of the the third derivative to find the real parts of the local maxima and minima. For each of those real maxima and minima, we use the fourth derivative to determine if they represent maxima (negative sign), or minima (positive sign). The function then finds the highest of any real local maxima. The values associated with the closest data point to highest real local maxima are taken as the threshold.
}
\examples{
# TODO write an example

}
\references{
Leo, J. A., Sabapathy, S., Simmonds, M. J., & Cross, T. J. (2017). The Respiratory Compensation Point is Not a Valid Surrogate for Critical Power. Medicine and science in sports and exercise, 49(7), 1452-1460.
Ozkaya, O., Balci, G. A., As, H., Cabuk, R., & Norouzi, M. (2022). Grey zone: a gap between heavy and severe exercise domain. Journal of Strength and Conditioning Research, 36(1), 113-120.
Wisén, A. G., & Wohlfart, B. (2004). A refined technique for determining the respiratory gas exchange responses to anaerobic metabolism during progressive exercise–repeatability in a group of healthy men. Clinical physiology and functional imaging, 24(1), 1-9.
Wisén, A. G., & Wohlfart, B. (2004). Aerobic and functional capacity in a group of healthy women: reference values and repeatability. Clinical physiology and functional imaging, 24(6), 341-351.
}
