% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/d2_reg_spline_maxima.R
\name{d2_reg_spline_maxima}
\alias{d2_reg_spline_maxima}
\title{Find a breakpoint using the maxima in the 2nd derivative of a regression spline}
\usage{
d2_reg_spline_maxima(
  .data,
  .x,
  .y,
  bp,
  ...,
  degree = 5,
  df = NULL,
  vo2 = "vo2",
  vco2 = "vco2",
  ve = "ve",
  time = "time",
  alpha_linearity = 0.05,
  pos_change = TRUE,
  ordering = c("by_x", "time")
)
}
\arguments{
\item{.data}{Gas exchange data frame or tibble.}

\item{.x}{The x-axis variable.}

\item{.y}{the y-axis variable.}

\item{bp}{Is this the first (\code{vt1}) or the second (\code{vt2}) breakpoint?}

\item{...}{Dot dot dot mostly allows this function to work properly if breakpoint() passes arguments that is not strictly needed by this function.

#' @references
Leo, J. A., Sabapathy, S., Simmonds, M. J., & Cross, T. J. (2017). The Respiratory Compensation Point is Not a Valid Surrogate for Critical Power. Medicine and science in sports and exercise, 49(7), 1452-1460.
Sherrill, D. L., Anderson, S. J., & Swanson, G. E. O. R. G. E. (1990). Using smoothing splines for detecting ventilatory thresholds. Medicine and Science in Sports and Exercise, 22(5), 684-689.
Wade, T. D., Anderson, S. J., Bondy, J., Ramadevi, V. A., Jones, R. H., & Swanson, G. D. (1988). Using smoothing splines to make inferences about the shape of gas-exchange curves. Computers and biomedical research, 21(1), 16-26.}

\item{degree}{The polynomial degree the function should fit to the curve. Most previous studies using this method use 3 or 5.}

\item{df}{If the user wants to use a specific number of knots, your `df` should equal the number of desired knots +`degree` to account for the df "consumed" by the polynomial regressions.}

\item{vo2}{Name of the \code{vo2} variable}

\item{vco2}{Name of the \code{vco2} variable}

\item{ve}{Name of the \code{ve} variable}

\item{time}{Name of the \code{time} variable}

\item{alpha_linearity}{Significance value to determine if a piecewise model explains significantly reduces the residual sums of squares more than a simpler model.}

\item{pos_change}{Do you expect the slope to be increasing at the breakpoint? This helps with filtering maxima.}

\item{ordering}{}
}
\value{
A slice of the original data frame at the threshold index with a new `algorithm` column.
}
\description{
#' Use regression splines of with an increasing number of knots to find the best-fit line for the data, take the second derivative, and find the highest maxima of the second derivative. This method is traditionally used with ventilatory equivalents (VE/VO2 or VE/VCO2), PetO2, or PetCO2 vs. time or vs. VO2.
}
\details{
#' Several works use an algorithm similar to this one. However, it is somewhat unclear if those previous works were truly using "smoothing splines" or "regression splines" (they are *not* the same). Previous works may refer to a third or fifth order "polynomial smoothing splines", which likely means "regression splines", but we cannot be certain at this time.

Unlike several published works, this method iteratively finds the best-fit regression by increasing the number of knots and using the likelihood ratio test with the \code{anova} function. When the likelihood ratio test does *not* find a statistically significant difference between the previous and the newest regression spline, this returns the previous regression spline. We found that this is a more conservative approach and yields fewer possible second derivative maxima to choose between. However, users can specify their desired number of knots.
}
\examples{
# TODO write an example

}
