% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outliers.R
\name{exercise_outliers}
\alias{exercise_outliers}
\title{Identify and remove outliers from exercise testing data}
\usage{
exercise_outliers(
  df,
  method = c("knn"),
  vars = c("time", "speed", "grade", "vo2_abs", "vco2", "ve"),
  k = 5,
  cutoff = 95,
  passes,
  dist_method = "euclidean",
  p = 1.5,
  keep_outliers = TRUE
)
}
\arguments{
\item{df}{Gas exchange data}

\item{method}{Outlier identification method. Choose between KNN or SD prediction interval.}

\item{vars}{Variables names to include in the distance matrix for KNN}

\item{k}{How many neighbors for KNN}

\item{cutoff}{The quantile of data points to keep with KNN or the SD prediction interval level.}

\item{passes}{How many times to loop through the KNN outlier method. If ]}

\item{dist_method}{The distance measure to be used. This must be one of "euclidean", "maximum", "manhattan", "canberra", "binary" or "minkowski". Any unambiguous substring can be given. See stats::dist.}

\item{p}{The power of the Minkowski distance. See stats::dist.}

\item{keep_outliers}{Should the function only identify outliers (\code{TRUE}) or remove them (\code{FALSE})?}
}
\description{
Identify and remove outliers from exercise testing data
}
\details{
The knn method normalizes the values that are passed to the distance matrix before calculating the distance matrix. They are normalized via \code{(.x - min(.x)) / (max(.x) - min(.x))}.
}
\examples{
# Add later
}
