% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_threshold_values.R
\name{find_threshold_vals}
\alias{find_threshold_vals}
\title{Find Other Variables at Ventilatory Thresholds}
\usage{
find_threshold_vals(
  .data,
  thr_x,
  thr_y,
  .x = .x,
  .y = .y,
  thr_calc_method = c("inv_dist", "dist_x_y", "dist_x"),
  k = 5,
  ...
)
}
\arguments{
\item{.data}{Gas exchange data frame or tibble.}

\item{thr_x}{X coordinate at the threshold.}

\item{thr_y}{Y coordinate at the threshold.}

\item{.x}{Name of the x variable used to find the threshold}

\item{.y}{Name of the y variable used to find the threshold}

\item{thr_calc_method}{Name of the threshold calculation method. Default is `inv_dist`. The next recommended method is `dist_x_y`, but `dist_x` is also an option.}

\item{k}{How many of the nearest data points should be used to calculate the inverse distance weights?}

\item{...}{Dot dot dot so this function plays nice with other functions that call it.}
}
\value{
A tibble
}
\description{
Choose between inverse distance weighted k-nearest neighbors interpolation (`inv-dist`, (default), `dist_x-y`, or `dist_x` to obtain the other values at ventilatory thresholds besides those found directly by the algorithm.
}
\details{
#' Most algorithms that identify ventilatory thresholds are NOT constrained to identify the threshold at an existing data point. Instead, most algorithms locate a point in space that *very* rarely coincides exactly with an existing data point. Given that, what are the values at the threshold *besides* the x and y coordinates you just found? The simplest approach is to find the closest observed data point in one dimension, usually x, and extract all the other variables at that observed data point.

The disadvantage of using the x-axis alone to determine the values at the threshold is that the closest observed data point in the x-dimension is often *not* the closest data point to the x-y coordinates of the threshold value. Using values from the closest-x data point may therefore produce misleading values. A better procedure finds the distance (Euclidean) from the threshold point to all observed data points. The nearest data point in the x and y dimensions thus provides the threshold data. The rationale is that an observed data point that is similar to the threshold in two dimensions is likely also more representative in the other dimensions that were not used to find the threshold.

However, what if the x-y coordinate of the threshold is close to several observed data points? Wouldn't an average be even better? The default method borrows from geospatial research and uses a version of inverse distanced weighted k-nearest neighbors interpolation. In geospatial research, given a data set of latitude, longitude, and elevation data, one can use this approach to predict the elevation at a new latitude and longitude value. The nearest elevations at known latitude and longitude values contribute more to the predicted elevation because "everything is related to everything else, but near things are more related than distant things" (Tobler, 1970).

In a similar fashion, this inverse-distance weight function calculates the Euclidean distance between all observed data points and the x-y threshold coordinate. Then, it selects from the nearest `k` points, defined by the user. Then, a "weight" is assigned proportional to the inverse of the distance of each point divided by the farthest distance of the farthest point ((d / max(d)^-1). The sum of the weights multiplied by their respective observed values become the values for each variable at the threshold.
}
\examples{

# TODO write an example

}
\references{
Tobler, W. R. (1970). A computer movie simulating urban growth in the Detroit region. Economic geography, 46(sup1), 234-240.
}
