% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interpolate.R
\name{interpolate}
\alias{interpolate}
\title{Interpolate graded exercise test data with linear or cubic spline interpolation}
\usage{
interpolate(.data, time_col, method = "linear", every_s = 1)
}
\arguments{
\item{.data}{Data frame to interpolate}

\item{time_col}{Time column from data frame to reference for interpolation}

\item{method}{Use linear or cubic spline interpolation}

\item{every_s}{Should the data be interpolated every second? Every 2 seconds?}
}
\value{
A data frame or tibble with numeric values interpolated to the specified time frame and by linear or cubic interpolation.
}
\description{
Interpolate graded exercise test data with linear or cubic spline interpolation
}
\details{
The time column must be in seconds.
}
\examples{
time <- sample.int(100, 20)
time <- sort(time)
y <- rnorm(length(time), mean = mean(time), sd = sd(time))
df <- data.frame(time = time, y = y)
interpolate(.data = df, time_col = "time", every_s = 2, method = "linear")
}
