% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intersection_point.R
\name{intersection_point}
\alias{intersection_point}
\title{Intersection point between two lines}
\usage{
intersection_point(lm1, lm2)
}
\arguments{
\item{lm1}{An \code{lm} ojbect.}

\item{lm2}{A second \code{lm} ojbect.}
}
\value{
Returns a named vector of point \code{(x, y)}
}
\description{
This function takes two \code{lm} objects in the form of \code{y = mx + b} and returns a vector of their \code{(x, y)} intersection point.
}
