% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/piecewise_helpers.R
\name{piecewise_stats}
\alias{piecewise_stats}
\title{Statistics of Piecewise Regressions Compared to a Simple Linear Regression}
\usage{
piecewise_stats(lm_left, lm_right, lm_simple)
}
\description{
Statistics of Piecewise Regressions Compared to a Simple Linear Regression
}
\keyword{internal}
