% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/start_end_test.R
\name{start_end_test}
\alias{start_end_test}
\title{Delete any 'pre-exercise' and 'recovery' portions from a graded exercise test}
\usage{
start_end_test(.data, intensity_col, start_intensity, end_intensity = 0)
}
\arguments{
\item{.data}{The data frame of the graded exercise test}

\item{intensity_col}{Speed or wattage values that help determine start and end of a test}

\item{start_intensity}{Initial intensity just after 'pre-exercise' data}

\item{end_intensity}{Intensity that denotes the test has ended}
}
\value{

}
\description{
Delete any 'pre-exercise' and 'recovery' portions from a graded exercise test
}
\examples{
df <- data.frame(time = c(14, 18, 61, 78, 88, 100, 120, 150, 220, 231),
speed = c(0, 0, 0, 0, 3, 3, 3, 7.4, 7.4, 0))
start_end_test(.data = df, intensity_col = "speed", start_intensity = 3)
}
