% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tr_mean.R
\name{tr_mean}
\alias{tr_mean}
\title{Trimmed mean helper function to emulate the 'mid 5 of 7' and similar averaging methods offered by MGC and Breeze}
\usage{
tr_mean(.v, x, y)
}
\arguments{
\item{.v}{A vector of numeric or logical values}

\item{x}{The number of values to keep}

\item{y}{The total number of values to consider}
}
\value{

}
\description{
Trimmed mean helper function to emulate the 'mid 5 of 7' and similar averaging methods offered by MGC and Breeze
}
\details{
Function for internal use only. It will not be exported.

This function trims the mean according to the difference between the x and y parameters. For example, if \code{x = 5} and \code{y = 7}, the difference is\code{2}. Half of the difference value is removed from the highest side of the vector and half the difference is removed from the lowest side of the vector. Therefore, the higest and lowest values would be removed before taking the mean. If instead \code{x = 5} and \code{y = 7}, the two lowest and two highest values would be removed before taking the mean.
}
\examples{

set.seed(2357234)
v <- 1:7
v <- round(v + rnorm(n = length(v), sd = 3), 2)
v
tr_mean(v, x = 5, y = 7)

}
\keyword{internal}
