% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outliers.R
\name{ventilatory_outliers}
\alias{ventilatory_outliers}
\title{Identify and remove outliers from exercise testing data}
\usage{
ventilatory_outliers(
  .data,
  outlier_cols = "vo2",
  time = "time",
  sd_lim = 3,
  width = 5,
  mos = "mean",
  align = "center",
  use_global_sd = TRUE,
  global_sd_mos = "median",
  exclude_test_val = TRUE,
  remove_outliers = TRUE,
  max_passes = Inf,
  plot_outliers = FALSE
)
}
\arguments{
\item{.data}{Gas exchange data frame or tibble.}

\item{outlier_cols}{Which columns should be used to find outliers. Default is \code{vo2} based on prevalent methodology.}

\item{time}{Name of the time column. Only used if plotting outliers}

\item{sd_lim}{How wide should the outlier interval be? Lamarra et al. (1987) originally suggested 3 standard deviations, but 4 is slightly more common.}

\item{width}{How many values should be considered when calculating the rolling window? Odd numbers are recommended.}

\item{mos}{Measure of center. Default is the \code{mean}.}

\item{align}{Should the rolling window be center, left, or right aligned? Default is \code{center}.}

\item{use_global_sd}{Calculate the rolling window using the global or local standard deviation? Local standard deviations can fluctuate in width considerably, so default is \code{TRUE}.}

\item{global_sd_mos}{Measure of center used to calculate the global standard deviation, if using. Default is \code{median}.}

\item{exclude_test_val}{Should the potential outlier be used or excluded from the rolling window? Default is to exclude (\code{TRUE}), because if an outlier affects the rolling window, it may not be appropriately flagged as an outlier.}

\item{remove_outliers}{Should the function return a data frame with outliers removed or merely noted in a new column?}

\item{max_passes}{By default, this function repeats itself (\code{max_passes = Inf}) until all values fit inside the rollowing window. However, users may wish to allow only a certain number of passes.}

\item{plot_outliers}{Plot outliers every iteration of the filter? Default is \code{FALSE}.}
}
\value{
A data frame or tibble with the rows containing outliers removed or retained according to the `remove_outliers` parameter. If outliers are removed, then the data frame returns a new column named `outliers`.
}
\description{
The current implementation only offers rolling-breath averages.
}
\examples{

# TODO add an example

}
\references{
Lamarra, N., Whipp, B. J., Ward, S. A., & Wasserman, K. (1987). Effect of interbreath fluctuations on characterizing exercise gas exchange kinetics. Journal of Applied Physiology, 62(5), 2003-2012.
}
