% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vo2_max.R
\name{vo2_plateau}
\alias{vo2_plateau}
\title{Determine if a graded exercise test demonstrated a VO2 plateau. Should there be an "all" test?}
\usage{
vo2_plateau(
  .data,
  test,
  method = "time",
  vo2_col = "vo2",
  time_col = "time",
  units = "mL",
  last_x = 30,
  delta_vo2 = 150,
  alpha = 0.05
)
}
\arguments{
\item{.data}{Gas exchange .data}

\item{test}{Test(s) to determine the presence of a plateau. Choose from slope at end of test (\code{slope_eot}), slope not significantly different from zero (\code{zero_slope}), and \code{vo2max_neighbor}. Supply a character vector (e.g. \code{c("slope_eot", "zero_slope")}) to run multiple tests at once and compare whether they agree; each test contributes one row to the output.}

\item{method}{Choose between \code{time} (default) or \code{breath}. Only used by the \code{slope_eot} and \code{zero_slope} tests, to determine how this function finds the end of the test: does \code{last_x} refer to the last \code{last_x} seconds, or the last \code{last_x} rows (breaths)? Most published methods (e.g. Yoon et al., 2007) use time; Myers et al. (1989, 1990) used breaths. \code{method = "breath"} assumes each row of \code{.data} represents one breath and issues a warning to that effect.}

\item{vo2_col}{The name of the VO2 column to use in .data. Can be relative or absolute. Make sure units align with \code{units}. The default is 150 mL/min, but other papers use 50, 100, 200, and even 280 mL/min (Midgley et al., 2007).}

\item{time_col}{The name of the time column in .data. This function assumes your time is in seconds.}

\item{units}{Specify if the O2 part of the VO2 units are in mL or in L}

\item{last_x}{How many seconds (\code{method = "time"}) or breaths (\code{method = "breath"}) at the end of test will be included for the slope_eot and zero_slope tests.}

\item{delta_vo2}{What is the acceptable variation in VO2 at the end of the test to consider if someone demonstrated a VO2 plateau.}

\item{alpha}{= The level of significance to use when applying the \code{zero_slope} method.
TODO add a plot function? This could highlight the VO2max point and locate the closest
# neighboring data point. For slope_eot it could add a regression line to the last_x
# of the graph.}
}
\value{
A tibble with a `test` column, a `plateau` column denoting if there was a plateau by the chosen test, and additional columns specific to each test. One row is returned per test requested; when tests with different output columns are combined, unused cells are filled with `NA`.
}
\description{
Determine if a graded exercise test demonstrated a VO2 plateau. Should there be an "all" test?
}
\details{
See Yoon et al. (2007) for a description of the \code{slope_eot} test. Briefly, if VO2-time slope in the last 30-60 seconds of exercise is less than a specified value, the participant demonstrated a plateau. Yoon et al. (2007) used a conservative value of 50 mL/min. Most papers appear to use 150 mL/min per Taylor et al. (1955). This test is likely more suitable when using rolling averages or digital filtering.

See Myers et al. (1989) and Myers et al. (1990) for a description of the \code{zero_slope} test. In short, this test tests if the slope is not significantly different from zero. If it was not significantly different from zero, the participant demonstrated a plateau. In this code, we also consider a VO2 plateau to be present if the slope was significantly different from zero while the sign is negative. In our personal experience, some subjects will demonstrate a plateau or a peak followed by a decrease in VO2 within the last 1-2 minutes of the test.
Technically this implementation is different from the Myers et al. studies. They used an 8-breath rolling average of the entire test and then performed a rolling 30-point regression of those 8-breath rolling average of VO2 vs. time. Once they had the slopes, they regressed that vs. time and determined which points were statistically significantly different from 0. To reiterate this point, they performed TWO regressions. First: VO2 vs. time using the 30 eight-breath, right-aligned rolling averages. This gave them the slope of VO2. They are repeatedly calculating the slope of the 8-breath right-aligned rolling averages across 30 of such data points. Then, they regressed this slope of VO2 against time (again). One question I have is: how much test time does a single one of these data points include? Recommendations center on calculating VO2max with no more than 30 seconds of data. If these 30-point slopes near the end of the test "stretch back" too far into the test, they might include more than 30 seconds of data. As such, that may may be too much time and you would be more likely to observe a non-zero slope of VO2.

See Astorino (2009), Astorino et al. (2000), and Robergs (2001) for a description of the vo2max_neighbor test. This test finds VO2max and then looks at the closest neighboring data point. If the difference between those two points is less than the specified \code{delta_VO2}, the participant demonstrated a plateau. This test is likely better suited when using bin averaging or in historical work using Douglas Bags when there were far fewer data points.
}
\examples{

# Load breath-by-breath graded exercise testing file
cpet_bbb <- utils::read.csv(
  system.file("extdata", "anton_vo2max_clean.csv", package = "gasExchangeR")
)

# VO2max neighbor test

# A 10-second bin average to showcase the VO2max neighbor test
cpet_10s_bin <- avg_exercise_test(
  cpet_bbb,
  test = "time",
  calc_type = "bin",
  time_col = "time",
  window = 10
)

vo2_plat_10s_bin <- vo2_plateau(cpet_10s_bin,
  test = "vo2max_neighbor",
  vo2_col = "vo2",
  time_col = "time",
  units = "mL",
  last_x = 60,
  delta_vo2 = 150,
  alpha = 0.05
)

vo2_plat_10s_bin

# A 30-second bin average to showcase the VO2max neighbor test
cpet_30s_bin <- avg_exercise_test(
  cpet_bbb,
  test = "time",
  calc_type = "bin",
  time_col = "time",
  window = 30
)

vo2_plat_30s_bin <- vo2_plateau(cpet_30s_bin,
  test = "vo2max_neighbor",
  vo2_col = "vo2",
  time_col = "time",
  units = "mL",
  last_x = 60,
  delta_vo2 = 150,
  alpha = 0.05
)

vo2_plat_30s_bin

# Notice how the presence of a VO2 plateau differs by a 10 (no plateau) vs.
# 30-second (yes plateau) bin average. This is due in part to the change in
# VO2 of 50 ml/min. Though opinions differ on the best bin average duration
# and VO2 differences for for determining VO2max and the presence of a VO2
# plateau, just be sure to clearly document your tests!

# Slope at end-of-test test

# We generally recommend a rolling breath average or digital filter for this
# approach as bin averages may result in too few data points.

cpet_15b_roll <- avg_exercise_test(
  cpet_bbb,
  test = "breath",
  calc_type = "rolling",
  time_col = "time",
  window = 15
)

vo2_plat_15b_roll_slope_eot <- vo2_plateau(cpet_15b_roll,
  test = "slope_eot",
  vo2_col = "vo2",
  time_col = "time",
  units = "mL",
  last_x = 60,
  delta_vo2 = 150,
  alpha = 0.05
)

vo2_plat_15b_roll_slope_eot

vo2_plat_15b_roll_zero_slope <- vo2_plateau(cpet_15b_roll,
 test = "zero_slope",
 vo2_col = "vo2",
 time_col = "time",
 units = "mL",
 last_x = 60,
 alpha = 0.05
)

vo2_plat_15b_roll_zero_slope

}
\references{
Astorino, T. A. (2009). Alterations in VO2 max and the VO2 plateau with manipulation of sampling interval. Clinical Physiology and Functional Imaging, 29(1), 60–67. https://doi.org/10.1111/j.1475-097X.2008.00835.x

Astorino, T. A., Robergs, R. A., Ghiasvand, F., Marks, D., & Burns, S. (2000). Incidence Of The Oxygen Plateau at VO2max During Exercise Testing To Volitional Fatigue. An International Electronic Journal, 3(4), 1–12. http://eprints.qut.edu.au/96933/1/96933.pdf

Midgley, A. W., McNaughton, L. R., Polman, R., & Marchant, D. (2007). Criteria for Determination of Maximal Oxygen Uptake: A Brief Critique and Recommendations for Future Research. Sports Medicine, 37(12), 1019–1028. https://doi.org/10.2165/00007256-200737120-00002

Midgley, A. W., Carroll, S., Marchant, D., McNaughton, L. R., & Siegler, J. (2009). Evaluation of true maximal oxygen uptake based on a novel set of standardized criteria. Applied Physiology, Nutrition, and Metabolism, 34(2), 115–123. https://doi.org/10.1139/H08-146

Myers, J., Walsh, D., Buchanan, N., & Froelicher, V. F. (1989). Can maximal cardiopulmonary capacity be recognized by a plateau in oxygen uptake? Chest, 96(6), 1312–1316. https://doi.org/10.1378/chest.96.6.1312

Myers, J., Walsh, D., Sullivan, M., & Froelicher, V. (1990). Effect of sampling on variability and plateau in oxygen uptake. Journal of Applied Physiology, 68(1), 404–410. https://doi.org/10.1152/jappl.1990.68.1.404

Robergs, R. A. (2001). An exercise physiologist’s “contemporary” interpretations of the “ugly and creaking edifices” of the VO2max concept. Journal of Exercise Physiology Online, 4(1), 1–44.

Taylor, H. L., Buskirk, E., & Henschel, A. (1955). Maximal Oxygen Intake as an Objective Measure of Cardio-Respiratory Performance. Journal of Applied Physiology, 8(1), 73–80. https://doi.org/10.1152/jappl.1955.8.1.73

Yoon, B. K., Kravitz, L., & Robergs, R. (2007). VO2max, protocol duration, and the VO2 plateau. Medicine and Science in Sports and Exercise, 39(7), 1186–1192. https://doi.org/10.1249/mss.0b13e318054e304
}
