% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wb_rc.R
\name{wb_rc}
\alias{wb_rc}
\title{Find the Respiratory Compensation Point as Implemented by WinBreak 3.7}
\usage{
wb_rc(
  .data,
  .x,
  .y,
  bp,
  vo2 = "vo2",
  vco2 = "vco2",
  ve = "ve",
  time = "time",
  alpha_linearity = 0.05,
  front_trim_vt1 = 60,
  front_trim_vt2 = 60,
  pos_change = TRUE,
  min_pct_change = 0.15,
  ...
)
}
\arguments{
\item{.data}{Gas exchange data.}

\item{.x}{The x-axis variable.}

\item{.y}{the y-axis variable.}

\item{bp}{Is this algorithm being used to find vt1 or vt2?}

\item{vo2}{The name of the vo2 column in \code{.data}}

\item{vco2}{The name of the vco2 column in \code{.data}}

\item{ve}{The name of the ve column in \code{.data}}

\item{time}{The name of the time column in \code{.data}}

\item{alpha_linearity}{Significance value to determine if a piecewise model explains significantly reduces the residual sums of squares more than a simpler model.}

\item{front_trim_vt1}{How much data (in seconds) to remove from the beginning of the test prior to fitting any regressions. The original V-slope paper suggests 1 minute.}

\item{front_trim_vt2}{How much data (in seconds) to remove from the beginning of the test prior to fitting any regressions. The original V-slope paper suggests 1 minute.}

\item{pos_change}{Do you expect the change in slope to be positive (default) or negative? If a two-line regression explains significantly reduces the sum square error but the change in slope does not match the expected underlying physiology, the breakpoint will be classified as indeterminate.}

\item{min_pct_change}{The original V-slope method (Beaver et al., 1986) considers the respiratory compensation point as determinant "if the change in slope"...between the left and right regressions..."is greater than a preselected amount (15\% of the initial slope)." See details for the specific WinBreak implementation.}

\item{...}{Dot dot dot mostly allows this function to work properly if breakpoint() passes arguments that is not strictly needed by this function.}
}
\value{
A list including slice of the original data frame at the threshold index with new columns `algorithm`, `determinant_bp`, `pct_slope_change`, `f_stat`, and `p_val_f.` The list also includes the fitted values, the left and right sides of the piecewise regression, and a simple linear regression.
}
\description{
This is an adaptation of WinBreak 3.7's (Epistemic Mindworks, 2003) method to find the respiratory compensation point (RCP), also known as the second ventilatory threshold (VT2). Their method is an adaptation of the original V-slope method by Beaver et al. (1986).
}
\details{
This uses a method similar to Ekkekakis et al. (2004) that adjusted the Beaver et al. (1986) method. In the original V-slope method, the RCP is the *first* time is the considered the first time the percent change from the left to the right regression exceeds 15\% (Beaver et al., 1986). However, Ekkekakis et al. (2004) also required a significant departure from linearity *and* used the largest percent increase between slopes observed by all divisions of the data. The reason, they state, is "because using a fixed slope difference (such as the 15\% mentioned by Beaver et al.) occasionally resulted in untenable results (e.g. points below 50\% VO2max)."

We note that this improvement can still yield innacurate results based on sudden changes in the tails of the data. For example the right regression may only contain two data points (the minimum to form a line) with a rather steep and positive slope because the very last data point is higher than the penultimate data point. The percent change from the left to the right regression at the second-to-last data point may therefore be the largest, and the function will choose this data point, despite the fact that the best-fit solution that still had a percent increase in slope of > 15\% was at a much earlier division of the data. Therefore, our implementation of WinBreak 3.7's (Epistemic Mindworks, 2003) RC method finds the division of the data that yields the smallest residual sums of squares, is a significant departure from linearity (significant F-test against a single regression line), and has a percent increase in slope by at least the stated percent (default = 15\%).
}
\examples{

# TODO write an example

}
\references{
Beaver, W. L., Wasserman, K., & Whipp, B. J. (1986). A new method for detecting anaerobic threshold by gas exchange. Journal of Applied Physiology, 121(6), 2020–2027.
Ekkekakis, P., Lind, E., Hall, E. E., & Petruzzello, S. J. (2008). Do regression-based computer algorithms for determining the ventilatory threshold agree?. Journal of Sports Sciences, 26(9), 967-976.
Epistemic Mindworks. (2003). WinBreak User Guide. Epistemic Mindworks.
}
