% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/averages.R
\name{avg_exercise_test}
\alias{avg_exercise_test}
\title{Average gas exchange data from an exercise test}
\usage{
avg_exercise_test(
  .data,
  type = c("breath", "time", "digital_filter"),
  subtype = c("rolling", "bin", "combo"),
  rolling_window = 15,
  bins = 15,
  align = "center"
)
}
\arguments{
\item{.data}{Gas exchange data}

\item{type}{Choose between breath averages, time averages, or digital filtering}

\item{subtype}{Choose rolling, bin, or combo}

\item{rolling_window}{How many seconds or breaths to include if rolling}

\item{bins}{Bin size of breaths or time}

\item{align}{If using a rolling method, how to align the rolling average}
}
\value{

}
\description{
This function averages first by either breath, time, or digital filtering
If averaging by breath or time averages, it can also perform rolling or bin averages. Furthermore, you can specify if you want a whole or trimmed mean.
}
\examples{

# TODO write an example later
}
