% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/averages.R
\name{avg_exercise_test}
\alias{avg_exercise_test}
\title{Average gas exchange data from an exercise test}
\usage{
avg_exercise_test(
  .data,
  type = "breath",
  subtype = "rolling",
  roll_window = 15,
  bins = 15,
  align = "center",
  mos = "mean",
  trim = 0,
  cutoff = 0.04,
  fs = 1,
  order = 3
)
}
\arguments{
\item{.data}{Gas exchange data.}

\item{type}{Choose between breath averages, time averages, or digital filtering.}

\item{subtype}{Choose rolling, bin, or combo.}

\item{roll_window}{How many seconds or breaths to include if rolling.}

\item{bins}{Bin size of breaths or time.}

\item{align}{If using a rolling method, how to align the rolling average. Other choices include "left", and "right".}

\item{mos}{'Measure of center'. Choices include "mean" or "median"}

\item{trim}{Indicate i you want a trimmed mean. This is used to emulate MCG's "mid-5-o-7" averaging method.}

\item{cutoff}{The cutoff frequency in Hz. Only used by digital filter.}

\item{fs}{The sampling frequency in Hz. Only used by digital filter.}

\item{order}{The Butterworth low-pass filter order. Only used by digital filter.}
}
\value{

}
\description{
This function averages first by either breath, time, or digital filtering
If averaging by breath or time averages, it can also perform rolling or bin averages. Furthermore, you can specify if you want a whole or trimmed mean.
}
\examples{

# TODO write an example later b/c I was getting errors earlier despite getting the functions to work when they were in the global environment

}
