% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/breakpoint.R
\name{breakpoint}
\alias{breakpoint}
\title{Find breakpoints in gas exchange data}
\usage{
breakpoint(
  .data,
  method = NULL,
  algorithm_vt1 = NULL,
  x_vt1 = NULL,
  y_vt1 = NULL,
  algorithm_vt2 = NULL,
  x_vt2 = NULL,
  y_vt2 = NULL,
  alpha_linearity = 0.05,
  vo2 = "vo2",
  vco2 = "vco2",
  ve = "ve",
  time = "time",
  bps = "both",
  truncate = TRUE,
  pos_change = TRUE,
  ...
)
}
\arguments{
\item{.data}{Gas exchange data}

\item{method}{Specify \code{excess_co2}, \code{v-slope}, or \code{v-slope_simple} and this fills in the most common variables used for these methods for those not already specified.}

\item{algorithm_vt1}{Algorithm to find VT1/GET}

\item{x_vt1}{\code{x} variable to use to fine VT1/GET/aerobic threshold}

\item{y_vt1}{\code{y} variable to use to fine VT1/GET/aerobic threshold}

\item{algorithm_vt2}{Algorithm to find VT2/RC}

\item{x_vt2}{\code{x} variable to use to fine VT2/RC/anaerobic threshold}

\item{y_vt2}{\code{y} variable to use to fine VT2/RC/anaerobic threshold}

\item{alpha_linearity}{Significance value to determine if a piecewise model explains significantly reduces the residual sums of squares more than a simpler model.}

\item{vo2}{The name of the vo2 column in \code{.data}}

\item{vco2}{The name of the vco2 column in \code{.data}}

\item{ve}{The name of the ve column in \code{.data}}

\item{time}{The name of the time column in \code{.data}}

\item{bps}{Should the function find the breakpoints for VT1, TVT2, or both. Default is \code{both}.}

\item{truncate}{By default this function truncates the data frame at VT2 prior to finding VT2. Change truncate to \code{FALSE} to use the entire data frame when searching for VT1.}

\item{pos_change}{Do you expect the change in slope to be positive (default) or negative? If a two-line regression explains significantly reduces the sum square error but the change in slope does not match the expected underlying physiology, the breakpoint will be classified as indeterminate.}

\item{...}{}
}
\description{
Breakpoint allows the user to specify particular x-y relationships, algorithms, and criteria for determining if a breakpoint was determinant or indeterminate. It is important to understand the order of steps the function takes.
\enumerate{
\item The function preselects x and y variables for the first and sometimes second threshold if the user indicates they wish to use a well-known method. For example, the v-slope method uses VE vs. VCO2 for finding VT2 (RC), and VCO2 vs. VO2 for finding VT1 (GET).
\item Determine the breakpoint at VT2 and if it was determinant.
\item If determinant, truncate the data at VT2.
\item Determine the breakpoint at VT1 and if it was determinant.
\item Return results.
}
}
\details{
There are multiple ways to find a breakpoint. The first is by specifying a commonly used \code{method} and the appropriate \code{algorithm} and \code{x}, and \code{y}, variables will be selected for you. For example, entering 'v-slope' for \code{method} will pre-select 'v-slope' for \code{algorithm}, 'vo2' for \code{x_vt1}, and 'vco2' for \code{y_vt1}. If you enter a specific \code{method}, you can override the default \code{algorithm}, \code{x}, and \code{y}.

Since there aren't common names for every way to find a breakpoint, you can also leave \code{method} as \code{NULL} and specify the \code{algorithm}, \code{x}, and \code{y}.

This function uses the excess CO2 formula as described by Gaskill et al. (2001).
}
\section{Warning}{

It is strongly advised the user selects \emph{absolute VO2 of the same units as VCO2 (e.g. mL/min for both variables)} rather than relative VO2. Using relative VO2 will likely lead to incorrect results.
}

\examples{
# TODO write an an example

}
\references{
Gaskill, S. E., Ruby, B. C., Walker, A. V. A. J., Sanchez, O. A., Serfass, R. C., & Leon, A. S. (2001). Validity and reliability of combining three methods to determine ventilatory threshold. Medicine & Science in Sports & Exercise, 33(11), 1841–1848.
}
