% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dmax.R
\name{dmax}
\alias{dmax}
\title{Finding a breakpoint using Cheng's Dmax method.}
\usage{
dmax(
  .data,
  .x,
  .y,
  vo2 = "vo2",
  vco2 = "vco2",
  ve = "ve",
  time = "time",
  alpha_linearity = 0.05,
  bp
)
}
\arguments{
\item{.data}{Gas exchange data.}

\item{.x}{The x-axis variable.}

\item{.y}{the y-axis variable.}

\item{vo2}{The name of the vo2 column in \code{.data}}

\item{vco2}{The name of the vco2 column in \code{.data}}

\item{ve}{The name of the ve column in \code{.data}}

\item{time}{The name of the time column in \code{.data}}

\item{alpha_linearity}{Significance value to determine if a piecewise model explains significantly reduces the residual sums of squares more than a simpler model.}

\item{bp}{Is this algorithm being used to find vt1 or vt2?}
}
\value{

}
\description{
Finding a breakpoint using Cheng's Dmax method.
}
\examples{
# TODO write examples
}
\references{
Cheng, B., Kuipers, H., Snyder, A. C., Keizer, H. A., Jeukendrup, A., & Hesselink, M. (1992). A new approach for the determination of ventilatory and lactate thresholds. International Journal of Sports Medicine, 13(7), 518–522. https://doi.org/10.1055/s-2007-1021309
}
