% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dmax.R
\name{dmax}
\alias{dmax}
\title{Finding a breakpoint using Cheng's Dmax algorithm}
\usage{
dmax(
  .data,
  .x,
  .y,
  vo2 = "vo2",
  vco2 = "vco2",
  ve = "ve",
  time = "time",
  alpha_linearity = 0.05,
  bp,
  pos_change = TRUE
)
}
\arguments{
\item{.data}{Gas exchange data.}

\item{.x}{The x-axis variable.}

\item{.y}{the y-axis variable.}

\item{vo2}{The name of the vo2 column in \code{.data}}

\item{vco2}{The name of the vco2 column in \code{.data}}

\item{ve}{The name of the ve column in \code{.data}}

\item{time}{The name of the time column in \code{.data}}

\item{alpha_linearity}{Significance value to determine if a piecewise model explains significantly reduces the residual sums of squares more than a simpler model.}

\item{bp}{Is this algorithm being used to find vt1 or vt2?}

\item{pos_change}{Do you expect the change in slope to be positive (default) or negative? If a two-line regression explains significantly reduces the sum square error but the change in slope does not match the expected underlying physiology, the breakpoint will be classified as indeterminate.}
}
\description{
As decribed by Cheng (1992) with modifications for brevity, a third order curvilinear regressions is first calculated. Pilot work demonstrated that a third order regression produced higher correlations coefficients than first and second order regressions, with the use of greater than a third order regression not significantly increasing the correlation coefficient. Next, a straight line formed by the two end points in each curve was drawn. Then a formula for computing the distance from point to line was used to calculate the distance from each point on the curve to the straight line. The point yielding the maximal distance (Dmax) derived from the computation was taken as the threshold.
}
\examples{
# TODO write examples
}
\references{
Cheng, B., Kuipers, H., Snyder, A. C., Keizer, H. A., Jeukendrup, A., & Hesselink, M. (1992). A new approach for the determination of ventilatory and lactate thresholds. International Journal of Sports Medicine, 13(7), 518–522. https://doi.org/10.1055/s-2007-1021309
}
