% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jm.R
\name{jm}
\alias{jm}
\title{Finding a breakpoint using the Jones-Molitoris method.}
\usage{
jm(.data, .x, .y, alpha_linearity = 0.05, bp)
}
\arguments{
\item{.data}{Gas exchange data.}

\item{.x}{The x-axis variable.}

\item{.y}{the y-axis variable.}

\item{alpha_linearity}{Significance value to determine if a piecewise model explains significantly reduces the residual sums of squares more than a simpler model.}

\item{bp}{Is this algorithm being used to find vt1 or vt2?}
}
\value{

}
\description{
Finding a breakpoint using the Jones-Molitoris method.
}
\examples{
# TODO write examples
}
\references{
Jones, R. H., & Molitoris, B. A. (1984). A statistical method for determining the breakpoint of two lines. Analytical Biochemistry, 141(1), 287–290. https://doi.org/10.1016/0003-2697(84)90458-5
}
