% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/orr.R
\name{orr}
\alias{orr}
\title{Finding a breakpoint using Orr's 'bruteforce' method.}
\usage{
orr(.data, .x, .y, alpha_linearity = 0.05, bp)
}
\arguments{
\item{.data}{Gas exchange data.}

\item{.x}{The x-axis variable.}

\item{.y}{the y-axis variable.}

\item{alpha_linearity}{Significance value to determine if a piecewise model explains significantly reduces the residual sums of squares more than a simpler model.}

\item{bp}{Is this algorithm being used to find vt1 or vt2?}
}
\value{

}
\description{
Finding a breakpoint using Orr's 'bruteforce' method.
}
\examples{
# TODO write examples
}
\references{
Orr, G. W., Green, H. J., Hughson, R. L., & Bennett, G. W. (1982). A computer linear regression model to determine ventilatory anaerobic threshold. Journal of Applied Physiology Respiratory Environmental and Exercise Physiology, 52(5), 1349–1352. https://doi.org/10.1152/jappl.1982.52.5.1349
}
