% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/orr.R
\name{orr}
\alias{orr}
\title{Finding a breakpoint using Orr's 'bruteforce' method}
\usage{
orr(
  .data,
  algorithm,
  .x,
  .y,
  vo2 = "vo2",
  vco2 = "vco2",
  ve = "ve",
  time = "time"
)
}
\arguments{
\item{.data}{Gas exchange data.}

\item{algorithm}{The specific breakpoint algorithm you want to use to find a given threshold. Choices include \code{dmax}, \code{dmax_mod}, \code{jm}, \code{orr}, \code{v-slope}, \code{v_slope_simple}, and \code{splines}.}

\item{.x}{The x-axis variable}

\item{.y}{the y-axis variable}

\item{vo2}{The name of the \code{vo2} variable}

\item{vco2}{The name of the \code{vco2} variable}

\item{ve}{The name of the \code{ve} variable}

\item{time}{The name of the \code{time} variable}
}
\value{

}
\description{
Finding a breakpoint using Orr's 'bruteforce' method
}
\examples{
# TODO write examples
}
\references{
Orr, G. W., Green, H. J., Hughson, R. L., & Bennett, G. W. (1982). A computer linear regression model to determine ventilatory anaerobic threshold. Journal of Applied Physiology Respiratory Environmental and Exercise Physiology, 52(5), 1349–1352. https://doi.org/10.1152/jappl.1982.52.5.1349
}
