% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/orr.R
\name{orr}
\alias{orr}
\title{Finding a breakpoint using Orr's 'bruteforce' algorithm.}
\usage{
orr(
  .data,
  .x,
  .y,
  bp,
  vo2 = "vo2",
  vco2 = "vco2",
  ve = "ve",
  time = "time",
  alpha_linearity = 0.05,
  front_trim = 60,
  pos_change = TRUE,
  ...
)
}
\arguments{
\item{.data}{Gas exchange data.}

\item{.x}{The x-axis variable.}

\item{.y}{the y-axis variable.}

\item{bp}{Is this algorithm being used to find vt1 or vt2?}

\item{vo2}{The name of the vo2 column in \code{.data}}

\item{vco2}{The name of the vco2 column in \code{.data}}

\item{ve}{The name of the ve column in \code{.data}}

\item{time}{The name of the time column in \code{.data}}

\item{alpha_linearity}{Significance value to determine if a piecewise model explains significantly reduces the residual sums of squares more than a simpler model.}

\item{front_trim}{How much data (in seconds) to remove from the beginning of the test prior to fitting any regressions. The original V-slope paper suggests 1 minute.}

\item{pos_change}{Do you expect the change in slope to be positive (default) or negative? If a two-line regression explains significantly reduces the sum square error but the change in slope does not match the expected underlying physiology, the breakpoint will be classified as indeterminate.}

\item{...}{Dot dot dot mostly allows this function to work properly if breakpoint() passes arguments that is not strictly needed by this function.}
}
\value{
A list including slice of the original data frame at the threshold index with new columns `algorithm`, `determinant_bp`, `pct_slope_change`, `f_stat`, and `p_val_f.` The list also includes the fitted values, the left and right sides of the piecewise regression, and a simple linear regression.
}
\description{
Finding a breakpoint using Orr's 'bruteforce' algorithm.
}
\examples{
# TODO write examples
}
\references{
Orr, G. W., Green, H. J., Hughson, R. L., & Bennett, G. W. (1982). A computer linear regression model to determine ventilatory anaerobic threshold. Journal of Applied Physiology Respiratory Environmental and Exercise Physiology, 52(5), 1349–1352. https://doi.org/10.1152/jappl.1982.52.5.1349
}
