% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spline_bp.R
\name{spline_bp}
\alias{spline_bp}
\title{Finding a breakpoint by iteratively moving an indicator variable}
\usage{
spline_bp(.data, .x, .y, degree = 1, alpha_linearity = 0.05, bp)
}
\arguments{
\item{.data}{Gas exchange data.}

\item{.x}{The x-axis variable.}

\item{.y}{the y-axis variable.}

\item{degree}{The degree of polynomial spline to use. Default is 1 to mimic other methods.}

\item{alpha_linearity}{Significance value to determine if a piecewise model explains significantly reduces the residual sums of squares more than a simpler model.}

\item{bp}{Is this algorithm being used to find vt1 or vt2?}
}
\value{

}
\description{
Finding a breakpoint by iteratively moving an indicator variable
}
\examples{
# TODO write an example
}
