% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/v-slope.R
\name{v_slope}
\alias{v_slope}
\title{Finding a breakpoint using Beaver's V-slope algorithm}
\usage{
v_slope(
  .data,
  .x,
  .y,
  vo2 = "vo2",
  vco2 = "vco2",
  ve = "ve",
  time = "time",
  method = NULL,
  slope_change_lim = 0.1,
  left_slope_lim = 0.6,
  alpha_linearity = 0.05,
  bp
)
}
\arguments{
\item{.data}{Gas exchange data.}

\item{.x}{The x-axis variable.}

\item{.y}{the y-axis variable.}

\item{vo2}{The name of the vo2 column in \code{.data}}

\item{vco2}{The name of the vco2 column in \code{.data}}

\item{ve}{The name of the ve column in \code{.data}}

\item{time}{The name of the time column in \code{.data}}

\item{method}{Pass \code{"v-slope"} to the method argument to exactly reproduce the procedure from the original paper. See the Warnings section for details.}

\item{slope_change_lim}{The absolute amount by which the slope should change from the left to the right regression line. Default per Beaver (1986) is 0.1.}

\item{left_slope_lim}{The original paper requires that the left regression line have a slope of > \code{0.6}.}

\item{alpha_linearity}{Significance value to determine if a piecewise model explains significantly reduces the residual sums of squares more than a simpler model.}

\item{bp}{Is this algorithm being used to find vt1 or vt2?}
}
\value{

}
\description{
V-slope uses the v-slope \emph{algorithm} to find a breakpoint. Note, the v-slope algorithm is different from the v-slope \emph{method} as a whole because the latter also includes a specific set of data processing steps (see reference). In order to use the v-slope algorithm as originally described by Beaver et al. (1986), pass \code{"v-slope"} to the \code{method} argument in the \code{breakpoint} function. Without indicating the \code{method} as \code{"v-slope"}, the function returns the breakpoint associated with ratio of the highest distance between the single regression line to the intersection point of the two regression lines, to the mean square error of the single regression line. This can sometimes lead to odd behavior and it is therefore generally recommended to specify \code{"v-slope"} in the \code{method} argument of \code{breakpoint} when using this function. See the Warnings and Details sections for details.
}
\details{
If the \code{method} argument is set to \code{"v-slope"}, the function enters a while loop to satisfy the quality control criteria (see Warnings section for details). Otherwise, the function returns the breakpoint as described the initial description.
}
\section{Warning}{

Using the v-slope algorithm as originally described by Beaver et al. (1986) is only intended to be used with the VCO2 vs. VO2 relationship when both are absolute measures of the same units (e.g. mL/min). The quality control criteria for this algorithm are an the increase in slope from the left to right regression line of > 0.1 and the slope of the left regression > 0.6. Using other x-y relationship leads to odd behavior that may not satisfy those criteria.
}

\examples{
# TODO write an example
}
\references{
Beaver, W. L., Wasserman, K., & Whipp, B. J. (1986). A new method for detecting anaerobic threshold by gas exchange. Journal of Applied Physiology, 121(6), 2020–2027.
}
