% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/v-slope.R
\name{v_slope}
\alias{v_slope}
\title{Finding a breakpoint using Beaver's V-slope algorithm}
\usage{
v_slope(
  .data,
  .x,
  .y,
  slope_change_lim = 0.1,
  vo2 = "vo2",
  vco2 = "vco2",
  ve = "ve",
  time = "time"
)
}
\arguments{
\item{.data}{Gas exchange data.}

\item{.x}{The x-axis variable.}

\item{.y}{the y-axis variable.}

\item{slope_change_lim}{The absolute amount by which the slope should change from the left to the right regression line. Default per Beaver (1986) is 0.1.}

\item{vo2}{The name of the \code{vo2} variable.}

\item{vco2}{The name of the \code{vco2} variable.}

\item{ve}{The name of the \code{ve} variable.}

\item{time}{The name of the \code{time} variable.}
}
\value{

}
\description{
Finding a breakpoint using Beaver's V-slope algorithm
}
\examples{
# TODO write an example
}
\references{
Beaver, W. L., Wasserman, K., & Whipp, B. J. (1986). A new method for detecting anaerobic threshold by gas exchange. Journal of Applied Physiology, 121(6), 2020–2027.
}
