% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/v-slope.R
\name{v_slope}
\alias{v_slope}
\title{Finding a breakpoint using Beaver's V-slope algorithm}
\usage{
v_slope(
  .data,
  .x,
  .y,
  bp,
  ...,
  vo2 = "vo2",
  vco2 = "vco2",
  ve = "ve",
  time = "time",
  ordering = c("by_x", "time"),
  method = NULL,
  slope_change_lim = 0.1,
  left_slope_lim = 0.6,
  front_trim_vt1 = 60,
  front_trim_vt2 = 60,
  alpha_linearity = 0.05,
  pos_change = TRUE,
  pos_slope_after_bp = TRUE,
  ci = FALSE,
  conf_level = 0.95,
  plots = TRUE
)
}
\arguments{
\item{.data}{Gas exchange data.}

\item{.x}{The x-axis variable.}

\item{.y}{the y-axis variable.}

\item{bp}{Is this algorithm being used to find vt1 or vt2?}

\item{...}{Dot dot dot mostly allows this function to work properly if breakpoint() passes arguments that is not strictly needed by this function.}

\item{vo2}{The name of the vo2 column in \code{.data}}

\item{vco2}{The name of the vco2 column in \code{.data}}

\item{ve}{The name of the ve column in \code{.data}}

\item{time}{The name of the time column in \code{.data}}

\item{ordering}{Prior to fitting any functions, should the data be reordered by the x-axis variable or by time? Default is to use the current x-axis variable and use the time variable to break any ties.}

\item{method}{Pass \code{"v-slope"} to the method argument to exactly reproduce the procedure from the original paper. See the Warnings section for details.}

\item{slope_change_lim}{The absolute amount by which the slope should change from the left to the right regression line. Default per Beaver (1986) is 0.1.}

\item{left_slope_lim}{The original paper requires that the left regression line have a slope of > \code{0.6}.}

\item{front_trim_vt1}{How much data (in seconds) to remove from the beginning of the test prior to fitting any regressions. The original V-slope paper suggests 1 minute.}

\item{front_trim_vt2}{See front_trim_vt1. This is here so this this function plays nice with `set_front_trim()`. The v-slope method should be used for finding VT1 only, so leave this alone.}

\item{alpha_linearity}{Significance value to determine if a piecewise model explains significantly reduces the residual sums of squares more than a simpler model.}

\item{pos_change}{Do you expect the change in slope to be positive (default) or negative? If a two-line regression explains significantly reduces the sum square error but the change in slope does not match the expected underlying physiology, the breakpoint will be classified as indeterminate.}

\item{pos_slope_after_bp}{Should the slope after the breakpoint be positive? Default is `TRUE`. This catches cases when the percent change in slope is positive, but the second slope is still negative. Change to `FALSE` when PetCO2 is the y-axis variable.}

\item{ci}{Should the output include confidence interval data? Default is `FALSE`.}

\item{conf_level}{Confidence level to use if calculating confidence intervals.}

\item{plots}{Should this function generate plots? Set to `FALSE` to save time.}
}
\value{
A list including slice of the original data frame at the threshold index with new columns `algorithm`, `determinant_bp`, `pct_slope_change`, `f_stat`, and `p_val_f.` The list also includes the fitted values, the left and right sides of the piecewise regression, and a simple linear regression.
}
\description{
V-slope uses the v-slope \emph{algorithm} to find a breakpoint. Note, the v-slope algorithm is different from the v-slope \emph{method} as a whole because the latter also includes a specific set of data processing steps (Beaver et al., 1986). In order to use the v-slope algorithm as originally described by Beaver et al. (1986), pass \code{"v-slope"} to the \code{method} argument in the \code{breakpoint} function. Without indicating the \code{method} as \code{"v-slope"}, the function returns the breakpoint associated with ratio of the highest distance between the single regression line to the intersection point of the two regression lines, to the mean square error of the piecewise regression lines. This can sometimes lead to odd behavior and it is therefore generally recommended to specify \code{"v-slope"} in the \code{method} argument of \code{breakpoint} when using this function. See the Warnings and Details sections for details.
}
\details{
If the \code{method} argument is set to \code{"v-slope"}, the function enters a while loop to satisfy the quality control criteria (see Warnings section for details). Otherwise, the function returns the breakpoint as described the initial description.

According to the Beaver et al. (1986, p. 2023) paper, the "intersection between the two regression lines is the tentative AT point." This usually provides a reasonable solution, but in some cases, the intersection of the best-fit model lies beyond the range of the x-axis. Instead of returning that value, we report the solution with the highest ratio of distance between the intersection point of the left and right regressions and the single regression line to the mean square error of the piecewise regression that also satisfies the following criteria: 1) the change in slope from the left to the right regression matches the anticipated change, usually positive; 2) the slope of the right regression matches the anticipated slope, usually positive; 3) the p-value of the F-test based on the extra sums of squares principle is statistically significant; and 4) the x-coordinate of the intersection point is within the range of the observed x-values. This differs slightly from the original method, but provides potentially plausible answers when the alternative is a negative VO2 or a VO2 well beyond VO2max.
}
\section{Warning}{

Using the v-slope algorithm as originally described by Beaver et al. (1986) is only intended to be used with the VCO2 vs. VO2 relationship when both are absolute measures of the same units (e.g. mL/min). The quality control criteria for this algorithm are an the increase in slope from the left to right regression line of > 0.1 and the slope of the left regression > 0.6. Using other x-y relationship leads to odd behavior that may not satisfy those criteria.
}

\examples{
# TODO write an example
}
\references{
Beaver, W. L., Wasserman, K., & Whipp, B. J. (1986). A new method for detecting anaerobic threshold by gas exchange. Journal of Applied Physiology, 121(6), 2020–2027.
}
